requirements.txt
Pillow
tqdm
litemapy
numpy